feat(syncv7): remove unused benchmarking tools
