+++ syncdoc/content/changelog/bench/Cargo.toml
@@ -10,12 +10,10 @@ sqlite = []
 diesel = { version = "2.3.7", features = ["postgres"] }
 clap = { version = "4", features = ["derive"] }
 plotters = "0.3"
-uuid = { version = "1", features = ["v4", "v7"] }
 rand = "0.10"
 anyhow = "1"
 toml = "1.1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-chrono = "0.4"
 urlencoding = "2"
 md-5 = "0.10"
