--- syncdoc/content/changelog/bench/config.toml
@@ -1,143 +1,0 @@
-# Changelog Insert Performance Benchmark Configuration
-
-# Global settings
-batch_size = 10_000          # inserts to measure per (scenario, N) pair
-output_dir = "results"
-
-# Table sizes to test
-# n_values = [1_200_000, 3_400_000, 5_600_000, 7_800_000, 10_000_000]
-# n_values = [100_000, 1_200_000, 2_300_000, 3_400_000, 4_500_000, 5_600_000, 6_700_000, 7_800_000, 8_900_000, 10_000_000]
-n_values = [1_000_000, 2_000_000, 5_000_000, 10_000_000, 20_000_000, 50_000_000, 75_000_000, 100_000_000]
-
-# Postgres connection
-[pg]
-host = "localhost"
-port = 5432
-user = "postgres"
-password = "postgres"
-database = "changelog_bench"
-
-# ── NULL distribution profiles (probability of NULL, 0.0–1.0) ──
-# Used by Phase 4 scenarios. source_site_id stays at fixed 25%.
-
-[null_profiles.mostly_null]
-store_id = 0.9
-transfer_store_id = 0.9
-patient_id = 0.9
-
-[null_profiles.balanced]
-store_id = 0.5
-transfer_store_id = 0.5
-patient_id = 0.5
-
-[null_profiles.mostly_present]
-store_id = 0.1
-transfer_store_id = 0.1
-patient_id = 0.1
-
-# ── Phase 1: PG Config comparison (same indexes = v7, no partitioning) ──
-# Goal: isolate the impact of Postgres configuration on insert performance
-# Settings are applied via ALTER SYSTEM + pg_reload_conf() before each scenario
-
-[[scenarios]]
-name = "v7_default"
-phase = 1
-indexes = "v7"
-pg_config_file = "pg-configs/default.txt"
-
-[[scenarios]]
-name = "v7_moderate"
-phase = 1
-indexes = "v7"
-pg_config_file = "pg-configs/moderate.txt"
-
-[[scenarios]]
-name = "v7_write_optimised"
-phase = 1
-indexes = "v7"
-pg_config_file = "pg-configs/write-optimised.txt"
-
-# ── Phase 2: Index strategy comparison ──
-# Goal: isolate the impact of different index sets on insert performance
-
-[[scenarios]]
-name = "idx_pk_only"
-phase = 2
-indexes = "index-configs/pk_only.sql"
-
-[[scenarios]]
-name = "idx_v7"
-phase = 2
-indexes = "index-configs/v7.sql"
-
-[[scenarios]]
-name = "idx_v7_all_partial"
-phase = 2
-indexes = "index-configs/v7_all_partial.sql"
-
-# ── Phase 3: Partition strategy comparison (same indexes = v7, write-optimised PG config) ──
-# Goal: isolate the impact of different partitioning strategies on insert performance
-
-[[scenarios]]
-name = "range_cursor_100k"
-phase = 3
-indexes = "v7"
-pg_config_file = "pg-configs/write-optimised.txt"
-partition_file = "partition-configs/range_cursor_100k.sql"
-
-[[scenarios]]
-name = "hash_cursor_32"
-phase = 3
-indexes = "v7"
-pg_config_file = "pg-configs/write-optimised.txt"
-partition_file = "partition-configs/hash_cursor_32.sql"
-
-[[scenarios]]
-name = "list_table_name"
-phase = 3
-indexes = "v7"
-pg_config_file = "pg-configs/write-optimised.txt"
-partition_file = "partition-configs/list_table_name.sql"
-
-# ── Phase 4: NULL distribution impact on v7 indexes, no partitioning ──
-# Goal: compare v7 vs v7_all_partial as null percentages vary
-# Uses grouped loop: UPDATE once per (N, profile), then cycle index types
-
-# mostly_null — 90% NULL
-[[scenarios]]
-name = "null90_v7"
-phase = 4
-indexes = "index-configs/v7.sql"
-null_profile = "mostly_null"
-
-[[scenarios]]
-name = "null90_v7_all_partial"
-phase = 4
-indexes = "index-configs/v7_all_partial.sql"
-null_profile = "mostly_null"
-
-# balanced — 50% NULL (matches seed default, no UPDATE needed)
-[[scenarios]]
-name = "null50_v7"
-phase = 4
-indexes = "index-configs/v7.sql"
-null_profile = "balanced"
-
-[[scenarios]]
-name = "null50_v7_all_partial"
-phase = 4
-indexes = "index-configs/v7_all_partial.sql"
-null_profile = "balanced"
-
-# mostly_present — 10% NULL
-[[scenarios]]
-name = "null10_v7"
-phase = 4
-indexes = "index-configs/v7.sql"
-null_profile = "mostly_present"
-
-[[scenarios]]
-name = "null10_v7_all_partial"
-phase = 4
-indexes = "index-configs/v7_all_partial.sql"
-null_profile = "mostly_present"
