--- syncdoc/content/changelog/bench/pg-configs/write-optimised.txt
@@ -1,11 +1,0 @@
-# Postgres 17 - Write-optimised configuration
-# Same as moderate but with synchronous_commit off
-# Tradeoff: crash could lose last ~0.5s of committed inserts
-
-shared_buffers = 1GB
-wal_buffers = 64MB
-synchronous_commit = off
-max_wal_size = 4GB
-checkpoint_completion_target = 0.9
-work_mem = 4MB
-autovacuum = on
