--- syncdoc/content/changelog/bench/src/bench.rs
@@ -1,384 +1,0 @@
-use anyhow::{Context, Result};
-use diesel::prelude::*;
-use diesel::sql_query;
-use rand::distr::Bernoulli;
-use rand::RngExt;
-use serde::{Deserialize, Serialize};
-use std::time::{Duration, Instant};
-
-use crate::config::{NullProfile, ScenarioConfig};
-use crate::schema::TABLE_NAME_VALUES;
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BenchResult {
-    pub scenario_name: String,
-    pub phase: u8,
-    pub n: u64,
-    pub batch_size: usize,
-    pub stats: Stats,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub null_profile: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Stats {
-    pub p50_us: u64,
-    pub p95_us: u64,
-    pub p99_us: u64,
-    pub mean_us: u64,
-    pub min_us: u64,
-    pub max_us: u64,
-}
-
-pub fn compute_stats(latencies: &mut Vec<Duration>) -> Stats {
-    assert!(!latencies.is_empty(), "Cannot compute stats on empty data");
-
-    latencies.sort();
-    let len = latencies.len();
-
-    let p50_idx = len / 2;
-    let p95_idx = ((len as f64) * 0.95) as usize;
-    let p99_idx = ((len as f64) * 0.99) as usize;
-
-    // Clamp to valid range
-    let p95_idx = p95_idx.min(len - 1);
-    let p99_idx = p99_idx.min(len - 1);
-
-    let total: Duration = latencies.iter().sum();
-    let mean = total / len as u32;
-
-    Stats {
-        p50_us: latencies[p50_idx].as_micros() as u64,
-        p95_us: latencies[p95_idx].as_micros() as u64,
-        p99_us: latencies[p99_idx].as_micros() as u64,
-        mean_us: mean.as_micros() as u64,
-        min_us: latencies[0].as_micros() as u64,
-        max_us: latencies[len - 1].as_micros() as u64,
-    }
-}
-
-/// Generate the SQL for pre-populating the changelog table using generate_series.
-/// Returns (sql, batch_from, batch_to) tuples for batched insertion.
-pub fn prepopulate_sql(n: u64) -> Vec<(String, u64, u64)> {
-    let batch_size: u64 = 1_000_000;
-    let mut batches = Vec::new();
-
-    let mut remaining = n;
-    let mut offset: u64 = 0;
-
-    while remaining > 0 {
-        let this_batch = remaining.min(batch_size);
-        let from = offset + 1;
-        let to = offset + this_batch;
-
-        let sql = format!(
-            "INSERT INTO changelog (record_id, table_name, row_action, source_site_id, store_id, transfer_store_id, patient_id)
-SELECT
-    md5(g::text)::uuid,
-    (ARRAY[{enum_array}])
-        [1 + (g % {enum_count})],
-    CASE WHEN g % 20 = 0 THEN 'DELETE'::row_action_type ELSE 'UPSERT'::row_action_type END,
-    CASE WHEN g % 4 = 0 THEN (g % 100)::integer ELSE NULL END,
-    CASE WHEN g % 2 = 0 THEN md5((g+1)::text)::uuid ELSE NULL END,
-    CASE WHEN g % 2 = 0 THEN md5((g+2)::text)::uuid ELSE NULL END,
-    CASE WHEN g % 2 = 0 THEN md5((g+3)::text)::uuid ELSE NULL END
-FROM generate_series({from}, {to}) AS g;",
-            enum_array = TABLE_NAME_VALUES
-                .iter()
-                .map(|v| format!("'{}'", v))
-                .collect::<Vec<_>>()
-                .join(", "),
-            enum_count = TABLE_NAME_VALUES.len(),
-            from = from,
-            to = to,
-        );
-
-        batches.push((sql, from, to));
-        offset += this_batch;
-        remaining -= this_batch;
-    }
-
-    batches
-}
-
-/// Generate a random INSERT statement for a single changelog row.
-/// When `profile` is `None`, uses 50% NULL defaults for all 3 UUID columns.
-/// When `Some`, uses the profile's null percentages.
-pub fn generate_random_insert(profile: Option<&NullProfile>) -> String {
-    let mut rng = rand::rng();
-
-    let table_name = TABLE_NAME_VALUES[rng.random_range(0..TABLE_NAME_VALUES.len())];
-    let record_id = uuid::Uuid::now_v7().to_string();
-
-    let row_action = if rng.sample(Bernoulli::from_ratio(1, 20).unwrap()) {
-        "DELETE"
-    } else {
-        "UPSERT"
-    };
-
-    // source_site_id: always 25% populated (fixed)
-    let source_site = if rng.sample(Bernoulli::from_ratio(1, 4).unwrap()) {
-        format!("{}", rng.random_range(1..100))
-    } else {
-        "NULL".to_string()
-    };
-
-    let (store_null, transfer_null, patient_null) = match profile {
-        Some(p) => (p.store_id, p.transfer_store_id, p.patient_id),
-        None => (0.5, 0.5, 0.5),
-    };
-
-    let store_id = if rng.sample(Bernoulli::new(1.0 - store_null).unwrap()) {
-        format!("'{}'", uuid::Uuid::now_v7())
-    } else {
-        "NULL".to_string()
-    };
-
-    let transfer_store_id = if rng.sample(Bernoulli::new(1.0 - transfer_null).unwrap()) {
-        format!("'{}'", uuid::Uuid::now_v7())
-    } else {
-        "NULL".to_string()
-    };
-
-    let patient_id = if rng.sample(Bernoulli::new(1.0 - patient_null).unwrap()) {
-        format!("'{}'", uuid::Uuid::now_v7())
-    } else {
-        "NULL".to_string()
-    };
-
-    format!(
-        "INSERT INTO changelog (record_id, table_name, row_action, source_site_id, store_id, transfer_store_id, patient_id) \
-         VALUES ('{}', '{}', '{}', {}, {}, {}, {})",
-        record_id, table_name, row_action, source_site, store_id,
-        transfer_store_id, patient_id,
-    )
-}
-
-/// Pre-populate the changelog table with N rows.
-pub fn prepopulate(conn: &mut PgConnection, n: u64) -> Result<()> {
-    if n == 0 {
-        return Ok(());
-    }
-
-    let batches = prepopulate_sql(n);
-    let total_batches = batches.len();
-
-    for (i, (sql, from, to)) in batches.iter().enumerate() {
-        if (i + 1) % 10 == 0 || i == 0 || i == total_batches - 1 {
-            eprintln!(
-                "  Pre-populating: batch {}/{} (rows {}-{})",
-                i + 1,
-                total_batches,
-                from,
-                to
-            );
-        }
-        sql_query(sql)
-            .execute(conn)
-            .with_context(|| format!("Failed to pre-populate batch {}", i + 1))?;
-    }
-
-    // ANALYZE and CHECKPOINT are skipped here — the seed container is
-    // throwaway, and ANALYZE is run after restore in the benchmark step.
-
-    Ok(())
-}
-
-/// Redistribute null percentages on the 3 UUID columns for Phase 4.
-/// Runs a single full-table UPDATE before indexes are created.
-pub fn redistribute_nulls(conn: &mut PgConnection, profile: &NullProfile) -> Result<()> {
-    let sql = format!(
-        "UPDATE changelog SET
-            store_id = CASE WHEN random() < {} THEN NULL ELSE md5(cursor::text)::uuid END,
-            transfer_store_id = CASE WHEN random() < {} THEN NULL ELSE md5((cursor+2)::text)::uuid END,
-            patient_id = CASE WHEN random() < {} THEN NULL ELSE md5((cursor+3)::text)::uuid END;",
-        profile.store_id, profile.transfer_store_id, profile.patient_id,
-    );
-    sql_query(&sql)
-        .execute(conn)
-        .context("Failed to redistribute null percentages")?;
-    Ok(())
-}
-
-/// Run the measurement phase: insert batch_size rows one at a time, timing each.
-pub fn measure_inserts(
-    conn: &mut PgConnection,
-    batch_size: usize,
-    _scenario: &ScenarioConfig,
-    profile: Option<&NullProfile>,
-) -> Result<Vec<Duration>> {
-    let mut latencies = Vec::with_capacity(batch_size);
-
-    for i in 0..batch_size {
-        let insert_sql = generate_random_insert(profile);
-
-        let start = Instant::now();
-        sql_query(&insert_sql)
-            .execute(conn)
-            .with_context(|| format!("Failed measurement insert {}", i))?;
-        latencies.push(start.elapsed());
-
-        if (i + 1) % 1000 == 0 {
-            eprintln!("  Measured {}/{} inserts", i + 1, batch_size);
-        }
-    }
-
-    Ok(latencies)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::config::IndexSet;
-
-    #[test]
-    fn test_compute_stats() {
-        let mut latencies: Vec<Duration> = (1..=100).map(|i| Duration::from_micros(i)).collect();
-
-        let stats = compute_stats(&mut latencies);
-
-        // Sorted: [1, 2, ..., 100]. Index 50 = value 51, index 95 = value 96, index 99 = value 100
-        assert_eq!(stats.p50_us, 51);
-        assert_eq!(stats.p95_us, 96);
-        assert_eq!(stats.p99_us, 100);
-        assert_eq!(stats.min_us, 1);
-        assert_eq!(stats.max_us, 100);
-        assert_eq!(stats.mean_us, 50); // mean of 1..100 = 50.5, truncated to 50
-    }
-
-    #[test]
-    fn test_compute_stats_single_element() {
-        let mut latencies = vec![Duration::from_micros(42)];
-
-        let stats = compute_stats(&mut latencies);
-
-        assert_eq!(stats.p50_us, 42);
-        assert_eq!(stats.p95_us, 42);
-        assert_eq!(stats.p99_us, 42);
-        assert_eq!(stats.min_us, 42);
-        assert_eq!(stats.max_us, 42);
-        assert_eq!(stats.mean_us, 42);
-    }
-
-    #[test]
-    fn test_generate_random_insert() {
-        let sql = generate_random_insert(None);
-
-        assert!(sql.starts_with("INSERT INTO changelog"));
-        assert!(sql.contains("record_id"));
-        assert!(sql.contains("table_name"));
-        assert!(sql.contains("row_action"));
-        assert!(sql.contains("source_site_id"));
-        assert!(sql.contains("store_id"));
-        assert!(sql.contains("transfer_store_id"));
-        assert!(sql.contains("patient_id"));
-
-        // Should NOT contain old columns
-        assert!(!sql.contains("name_link_id"));
-        assert!(!sql.contains("is_sync_update"));
-
-        // Verify it contains a valid table_name enum value
-        let has_valid_table_name = TABLE_NAME_VALUES
-            .iter()
-            .any(|v| sql.contains(&format!("'{}'", v)));
-        assert!(has_valid_table_name, "SQL should contain a valid table_name enum value");
-
-        // Verify row_action is valid
-        assert!(sql.contains("'UPSERT'") || sql.contains("'DELETE'"));
-    }
-
-    #[test]
-    fn test_generate_random_insert_with_profile() {
-        let profile = NullProfile {
-            store_id: 0.9,
-            transfer_store_id: 0.9,
-            patient_id: 0.9,
-        };
-        let sql = generate_random_insert(Some(&profile));
-
-        assert!(sql.starts_with("INSERT INTO changelog"));
-        assert!(sql.contains("record_id"));
-        assert!(sql.contains("store_id"));
-    }
-
-    #[test]
-    fn test_prepopulate_sql_generation() {
-        let batches = prepopulate_sql(1_000_000);
-
-        // 1M / 1M = 1 batch
-        assert_eq!(batches.len(), 1);
-
-        // Single batch should be 1..1000000
-        assert!(batches[0].0.contains("generate_series(1, 1000000)"));
-        assert_eq!(batches[0].1, 1);
-        assert_eq!(batches[0].2, 1_000_000);
-
-        // Should NOT contain old columns
-        assert!(!batches[0].0.contains("name_link_id"));
-        assert!(!batches[0].0.contains("is_sync_update"));
-    }
-
-    #[test]
-    fn test_prepopulate_sql_zero() {
-        let batches = prepopulate_sql(0);
-        assert!(batches.is_empty());
-    }
-
-    #[test]
-    fn test_prepopulate_sql_uneven_batches() {
-        let batches = prepopulate_sql(2_500_000);
-
-        // 2.5M / 1M = 2 full batches + 1 partial
-        assert_eq!(batches.len(), 3);
-        assert!(batches[2].0.contains("generate_series(2000001, 2500000)"));
-    }
-
-    #[test]
-    #[ignore] // Requires a running Postgres
-    fn test_measurement_inserts() {
-        use crate::config::PgConfig;
-        use crate::db;
-        use crate::schema;
-
-        let pg = PgConfig::localhost("changelog_bench_test_measure");
-
-        db::reset_database(&pg).unwrap();
-        let mut conn = db::connect(&pg, std::time::Duration::from_secs(5)).unwrap();
-
-        let scenario = ScenarioConfig {
-            name: "test".to_string(),
-            phase: 1,
-            indexes: IndexSet::V7,
-            ..Default::default()
-        };
-
-        let stmts = schema::setup_sql(&scenario, 0, 100);
-        for stmt in &stmts {
-            sql_query(stmt).execute(&mut conn).unwrap();
-        }
-
-        let latencies = measure_inserts(&mut conn, 100, &scenario, None).unwrap();
-
-        assert_eq!(latencies.len(), 100);
-        assert!(latencies.iter().all(|d| d.as_nanos() > 0));
-
-        #[derive(diesel::QueryableByName)]
-        struct CountRow {
-            #[diesel(sql_type = diesel::sql_types::BigInt)]
-            cnt: i64,
-        }
-
-        let result: Vec<CountRow> =
-            sql_query("SELECT count(*)::bigint AS cnt FROM changelog")
-                .load(&mut conn)
-                .unwrap();
-        assert_eq!(result[0].cnt, 100);
-
-        // Cleanup
-        let maint_str = pg.maintenance_connection_string();
-        drop(conn);
-        let mut maint = diesel::PgConnection::establish(&maint_str).unwrap();
-        let _ = sql_query(&format!("DROP DATABASE IF EXISTS \"{}\"", pg.database))
-            .execute(&mut maint);
-    }
-}
