+++ syncdoc/content/changelog/bench/src/chart.rs
